backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
# tensorflow - removed due to incompatibility with python 3.14, using simulation mode
opencv-python-headless
numpy
scikit-learn
pandas
sqlmodel
sqlmodel
aiosqlite
google-generativeai
bcrypt
opencv-python-headless
numpy
onnx



